新年的第一天未完成的测试

+++ airport/general.h
@@ -7,7 +7,7 @@ struct passenger
     QString name;//成员姓名
     QString ID;//身份id
     QString bookair;//订票航班号
-    long long ticket;//订票量
+    int ticket;//订票量
     int level;//订票舱等级
 //    passenger(qstring a, long long b, int c) : name(a), ticket(b), level(c){}
 };
@@ -21,7 +21,7 @@ public:
     QString airnum;//航班号
     QString planenum;//飞机号
     QString day;//飞机周日
-    long long people;//乘员定额
-    long long retickt;//余票量
+    int people;//乘员定额
+    int retickt;//余票量
 };
 #endif // GENERAL_H
